move codes from routes/login to routes/apiv3/register
